SQUASHME: fix gatsby helper usage errors
